% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trait_model.R
\name{fit_trait_model}
\alias{fit_trait_model}
\title{Fits a Markov model of trait evolution along a phylogenetic tree}
\usage{
fit_trait_model(tree_trait)
}
\arguments{
\item{tree_trait}{A `phylo4d` object containing a phylogenetic tree and trait
data at each of the tips}
}
\value{
`fitMK` object from `phytools` package
}
\description{
Fits a Markov model of trait evolution along a phylogenetic tree
}
