% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_traits}
\alias{tree_traits}
\title{Simulated phylogenetic and trait data}
\format{
An object of class `phylo4d`.
}
\usage{
tree_traits
}
\description{
A dataset containing a phylogenetic tree and discrete trait data (states 1,
2 and 3) for each tip in the tree.
}
\keyword{datasets}
